Add files via upload
Use Python to Build Home Security Camera


Inspiration for this build:
The idea for this particular build started with a spate of package thefts in my neighborhood. I researched various WiFi enabled home security cameras, the average cost of which usually floated in the $100 to $200 range. Since I had a basic knowledge of Python already on hand, I decided to try building my own